+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2025-04-26 deimos 0.8.1, deimos_editor 0.1.0
+
+Implement clone workaround for `Box<dyn Calc>` via typetag's enum vtable.
+
+### Added
+
+* impl Clone for `Box<dyn Calc>` by sending for a roundtrip loop through serde_json
+  * Supports construction from prototype in GUI editor
+* Add `deimos_editor` node editor GUI for calc graph
+
+
 ## 2025-03-23 deimos 0.8.0
 
 Implement `SequenceMachine` calc, which provides arbitrary lookup-table state machines in order to support